🚚  move useProjectContributors to project utils
